Rename texture_cache to texture_node
