+++ src/conversion/mod.rs
@@ -2,4 +2,4 @@ pub mod light_shape;
 pub mod mesh_data;
 pub mod plane_data;
 pub mod spectrum;
-pub mod texture_cache;
+pub mod texture_node;
