+++ src/conversion/texture_node/create_image_variants.rs
