+++ src/conversion/texture_node/dyna_image.rs
