+++ src/conversion/texture_node/old/texture_cache_manager.rs
