+++ src/conversion/texture_node/render_texture_image.rs
